Created the basic structure for the Revision History plugin.
https://github.com/unfoldingWord-dev/Door43/issues/105

In order to inform our platforms that changes have been made to the
Door43 website, we need to develop a simple API to request recent
changes to the DokuWiki website.  This plugin will provide users a way
to retrieve the revision history for specific media
or page using an API GET Request.  This is the basic structure for a
DokuWiki plugin that will provide that access.