Created a simple API for retrieving revision history
https://github.com/unfoldingWord-dev/Door43/issues/105

Created a simple API for retrieving the revision history of any files or
media changed in the DokuWiki site.  This action plugin adds a
revhistory action where a user can receive a JSON file of the current
history.  This plugin provides the ability to filter that information
by ending date, namespace, and starting date.  The user can also specify
whether they just want file changes, or to include media changes.
Finally, the user can provide a parameter to sort the history in
ascending or descending order.

* The action 'revhistory' will trigger this plugin
* This plugin adds several configuration parameters in the Configuration
* Manager of DokuWiki