Show diff links for status-tagged pages
This plugin adds links to pages that point to diff reports.  These links
only appear on pages that use status tags (e.g. 'check', 'review',
'publish').  These diff reports show the difference between pages when they
most recently changed to 'check', 'review', or 'publish', as well as a diff
between the most recent change to 'publish' and the current revision.  If
the page has no status changes, the links do not appear.

Because the links must be dynamically generated every time the page is
displayed (because the status tags may have changed since the last time the
page was displayed), the plugin will force a cache reset for the main page
content.  It does not affect caching for other pages (e.g. sidebar and
other included pages).